+++ emulator/src/glfw/lib/cocoa/cocoa_window.m
@@ -27,7 +27,7 @@
 //
 //========================================================================
 
-#include <internal.h>
+#include "internal.h"
 
 #include <AvailabilityMacros.h>