+++ emulator/src/glfw/lib/win32/win32_init.c
@@ -28,7 +28,7 @@
 //
 //========================================================================
 
-#include <internal.h>
+#include "internal.h"
 
 // With the Borland C++ compiler, we want to disable FPU exceptions
 #ifdef __BORLANDC__