+++ emulator/src/glfw/lib/x11/x11_glext.c
@@ -28,7 +28,7 @@
 //
 //========================================================================
 
-#include <internal.h>
+#include "internal.h"
 
 
 void (*glXGetProcAddress(const GLubyte *procName))();